:memo: rename repository from provider to adapter
:truck: rename repository from provider to adapter

Renamed the MQSS Pennylane Provider -> MQSS Pennylane Adapter. Changed the folders and other imports accordingly.

:truck: rename repository from provider to adapter

:truck: rename repository from provider to adapter
