+++ test/test_backend.py
@@ -2,7 +2,7 @@ import pennylane as qml
 
 import pytest
 from src.config import MQSS_TOKEN, MQSS_BACKENDS
-from src.mqp.pennylane_provider.device import MQSSPennylaneDevice
+from src.mqp.pennylane_adapter.device import MQSSPennylaneDevice
 from pennylane import numpy as np
 
 dev = MQSSPennylaneDevice(wires=2, token=MQSS_TOKEN, backends=MQSS_BACKENDS)
